% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cluster_matrix}
\alias{cluster_matrix}
\title{Cluster-robust Standard Error Estimation}
\usage{
cluster_matrix(e, cluster, x)
}
\arguments{
\item{e}{Vector of residuals.}

\item{cluster}{Vector of clusters.}

\item{x}{Regressor matrix.}
}
\value{
Gives the XeeX matrix.
}
\description{
\code{cluster_matrix} is a helper for computation of cluster-robust standard errors.
}
