% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xvalidate.R
\name{compute_fes}
\alias{compute_fes}
\title{Fixed Effects Computation}
\usage{
compute_fes(
  y,
  fes,
  x,
  b,
  insample_obs = rep(1, n),
  onlymus = FALSE,
  tol = 1e-08,
  verbose = FALSE
)
}
\arguments{
\item{y}{Dependent variable (a vector).}

\item{fes}{List of fixed effects.}

\item{x}{Regressor matrix.}

\item{b}{TODO: check what this does.}

\item{insample_obs}{Numeric vector with selected observations (TODO: check this better).}

\item{onlymus}{Logical. If \code{TRUE}, returns only mus (TODO: explain this).}

\item{tol}{A tolerance parameter.}

\item{verbose}{Logical. If \code{TRUE}, prints messages to the console while evaluating.}
}
\value{
A list (TODO: check and explain better).
}
\description{
Compute FEs using PPML FOCs (TODO: explain this better?).
}
\examples{
#TODO: add some examples.
}
