% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{Country ISO Codes}
\format{
A data frame with 249 rows and 4 variables.
\itemize{
\item iso: Country ISO code.
\item name: Country name.
\item region: Continent.
\item subregion: sub-continental region.
}
}
\source{
https://github.com/lukes/ISO-3166-Countries-with-Regional-Codes.
}
\usage{
countries
}
\description{
An auxiliary data set with basic geographic information about ISO codes included in the \code{trade}
data set.
}
\keyword{datasets}
