% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastolsCpp}
\alias{fastolsCpp}
\title{Faster Least Squares Estimation}
\usage{
fastolsCpp(X, y)
}
\arguments{
\item{X}{Regressor matrix.}

\item{y}{Dependent variable (a vector).}
}
\value{
The vector of parameter (beta) estimates.
}
\description{
Finds Least Squares solutions using C++.
}
