% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{faststddev}
\alias{faststddev}
\title{Faster Standard Deviation}
\usage{
faststddev(X, w)
}
\arguments{
\item{X}{Regressor matrix.}

\item{w}{Weights.}
}
\value{
Vector of standard deviations of the parameter estimates.
}
\description{
Computes standard deviation using C++.
}
