% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastwmean}
\alias{fastwmean}
\title{Faster Weighted Mean}
\usage{
fastwmean(X, w)
}
\arguments{
\item{X}{Regressor matrix.}

\item{w}{Weights.}
}
\value{
Weighted mean.
}
\description{
Computes weighted mean using C++.
}
