% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdfeppml.R
\name{hdfeppml}
\alias{hdfeppml}
\title{Title}
\usage{
hdfeppml(
  y,
  x,
  fes,
  tol = 1e-08,
  hdfetol = 1e-04,
  colcheck = TRUE,
  selectobs = NULL,
  mu = NULL,
  saveX = TRUE,
  init_z = NULL,
  verbose = FALSE,
  maxiter = 1000,
  cluster = NULL,
  vcv = TRUE
)
}
\arguments{
\item{vcv}{}
}
\value{

}
\description{
Title
}
