% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceberg.R
\name{iceberg}
\alias{iceberg}
\title{Iceberg Lasso Implementation (in development)}
\usage{
iceberg(data, dep, indep = NULL, selectobs = NULL, ...)
}
\arguments{
\item{data}{A data frame containing all relevant variables.}

\item{dep}{A string with the name of the independent variable or a column number.}

\item{indep}{A vector with the names or column numbers of the regressors. If left unspecified,
all remaining variables (excluding fixed effects) are included in the regressor matrix.}

\item{selectobs}{Optional. A vector indicating which observations to use (either a logical vector
or a numeric vector with row numbers, as usual when subsetting in R).}

\item{...}{Further arguments, to be passed on to the main function.}
}
\value{
A matrix with coefficient estimates for all dependent variables.
}
\description{
TODO.
}
\examples{
\dontrun{iceberg_results <- iceberg(data = trade[, -(1:6)],
        dep = c("ad_prov_14", "cp_prov_23", "tbt_prov_07",
                "tbt_prov_33", "tf_prov_41", "tf_prov_45"),
        selectobs = (trade$time == "2016"))}
}
