% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{manyouter}
\alias{manyouter}
\title{Many Outer Products}
\usage{
manyouter(A, B, c)
}
\arguments{
\item{A, B}{Numeric vectors.}

\item{c}{Integer.}
}
\description{
Compute a large number of outer products (useful for clustered SEs) using C++.
}
