% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlfitppml.R
\name{mlfitppml}
\alias{mlfitppml}
\title{Title}
\usage{
mlfitppml(
  y,
  x,
  fes,
  lambdas,
  penalty = c("lasso", "ridge"),
  tol = 1e-08,
  hdfetol = 1e-04,
  colcheck = TRUE,
  post = TRUE,
  cluster = NULL,
  method = c("bic", "iterative"),
  IDs = 1:n,
  verbose = FALSE,
  xval = FALSE,
  standardize = TRUE,
  vcv = TRUE,
  penweights = NULL,
  K = 15
)
}
\arguments{
\item{K}{}
}
\value{

}
\description{
Title
}
