% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlfitppml_int.R
\name{mlfitppml_int}
\alias{mlfitppml_int}
\title{General Penalized PPML Estimation}
\usage{
mlfitppml_int(
  y,
  x,
  fes,
  lambdas,
  penalty = "lasso",
  tol = 1e-08,
  hdfetol = 1e-04,
  colcheck = TRUE,
  post = TRUE,
  cluster = NULL,
  method = "bic",
  IDs = 1:n,
  verbose = FALSE,
  xval = FALSE,
  standardize = TRUE,
  vcv = TRUE,
  penweights = NULL,
  K = 15
)
}
\arguments{
\item{y}{Dependent variable (a vector)}

\item{x}{Regressor matrix.}

\item{fes}{List of fixed effects.}

\item{lambdas}{Vector of penalty parameters.}

\item{penalty}{A string indicating the penalty type. Currently supported: "lasso" and "ridge".}

\item{tol}{Tolerance parameter for convergence of the IRLS algorithm.}

\item{hdfetol}{Tolerance parameter for the within-transformation step,
passed on to \code{lfe::demeanlist}.}

\item{colcheck}{Logical. If \code{TRUE}, checks for perfect multicollinearity in \code{x}.}

\item{post}{Logical. If \code{TRUE}, estimates a post-penalty model with the selected variables.}

\item{cluster}{Optional: a vector classifying observations into clusters (to use when calculating SEs).}

\item{method}{The user can set this equal to "plugin" to perform the plugin algorithm with
coefficient-specific penalty weights (see details). Otherwise, a single global penalty is used.}

\item{IDs}{TODO: check what this does (probably for cross validation).}

\item{verbose}{Logical. If \code{TRUE}, it prints information to the screen while evaluating.}

\item{xval}{Logical. If \code{TRUE}, carries out cross-validation.}

\item{standardize}{Logical. If \code{TRUE}, x variables are standardized before estimation.}

\item{vcv}{Logical. If \code{TRUE} (the default), the post-estimation model includes standard errors.}

\item{penweights}{Optional: a vector of coefficient-specific penalties to use in plugin lasso when
\code{method == "plugin"}.}

\item{K}{TODO: check what this does.}
}
\value{
A list (TODO: complete this).
}
\description{
\code{mlfitppml_int} is a general wrapper function for penalized PPML estimation, calling
\code{penhdfeppml_int}, \code{penhdfeppml_cluster} and \code{hdfeppml_int} as needed.
}
\examples{
# TODO: add examples here.
}
