% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penhdfeppml_cluster_int.R
\name{penhdfeppml_cluster_int}
\alias{penhdfeppml_cluster_int}
\title{Plugin Lasso Estimation}
\usage{
penhdfeppml_cluster_int(
  y,
  x,
  fes,
  cluster,
  tol = 1e-08,
  hdfetol = 1e-04,
  glmnettol = 1e-12,
  penalty = "lasso",
  penweights = NULL,
  saveX = TRUE,
  mu = NULL,
  colcheck = TRUE,
  K = 15,
  init_z = NULL,
  post = FALSE,
  verbose = FALSE,
  lambda = NULL
)
}
\arguments{
\item{y}{Dependent variable (a vector)}

\item{x}{Regressor matrix.}

\item{fes}{List of fixed effects.}

\item{cluster}{Optional: a vector classifying observations into clusters (to use when calculating SEs).}

\item{tol}{Tolerance parameter for convergence of the IRLS algorithm.}

\item{hdfetol}{Tolerance parameter for the within-transformation step,
passed on to \code{lfe::demeanlist}.}

\item{glmnettol}{Tolerance parameter to be passed on to \code{glmnet::glmnet}.}

\item{penalty}{A string indicating the penalty type. Currently supported: "lasso" and "ridge".}

\item{penweights}{Optional: a vector of coefficient-specific penalties to use in plugin lasso when
\code{method == "plugin"}.}

\item{saveX}{Logical. If \code{TRUE}, it returns the values of x and z after partialling out the
fixed effects.}

\item{mu}{Optional: initial values of the \eqn{\mu} "weights", to be used in the
first iteration of the algorithm.}

\item{colcheck}{Logical. If \code{TRUE}, checks for perfect multicollinearity in \code{x}.}

\item{K}{TODO: check what this does.}

\item{init_z}{Optional: initial values of the transformed dependent variable, to be used in the
first iteration of the algorithm.}

\item{post}{Logical. If \code{TRUE}, estimates a post-penalty model with the selected variables.}

\item{verbose}{Logical. If \code{TRUE}, it prints information to the screen while evaluating.}

\item{lambda}{Penalty parameter (a number).}
}
\value{
TODO: check this.
}
\description{
Estimates coefficient-specific penalty weights that account for heteroskedasticity.
Called by \code{mlfitppml_int} and \code{penhdfeppml_int}.
}
\examples{
# TODO: add examples here.
}
