% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin_lasso.R
\name{plugin_lasso}
\alias{plugin_lasso}
\title{Iceberg Lasso Implementation (in development)}
\usage{
plugin_lasso(data, dep = 1, indep = NULL, selectobs = NULL, ...)
}
\arguments{
\item{data}{A data frame containing all relevant variables.}

\item{dep}{A string with the name of the independent variable or a column number.}

\item{indep}{A vector with the names or column numbers of the regressors. If left unspecified,
all remaining variables (excluding fixed effects) are included in the regressor matrix.}

\item{selectobs}{Optional. A vector indicating which observations to use (either a logical vector
or a numeric vector with row numbers, as usual when subsetting in R).}

\item{...}{Further arguments, to be passed on to the main function.}
}
\value{
A list with (TODO).
}
\description{
TODO.
}
\examples{
# TODO
}
