% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceberg.R
\name{plugin_lasso_int}
\alias{plugin_lasso_int}
\title{Iceberg Lasso Implementation (in development)}
\usage{
plugin_lasso_int(
  y,
  x,
  tol = 1e-08,
  glmnettol = 1e-12,
  penweights = NULL,
  colcheck = FALSE,
  K = 50,
  verbose = FALSE,
  lambda = NULL,
  phipost = FALSE
)
}
\arguments{
\item{y}{Dependent variable (a vector).}

\item{x}{Regressor matrix.}

\item{tol}{Tolerance parameter.}

\item{glmnettol}{Tolerance parameter to be passed on to glmnet.}

\item{penweights}{TODO: check what this does (parameter-specific penalties in plugin lasso?).}

\item{colcheck}{Logical. If \code{TRUE}, checks for perfect multicollinearity in \code{x}.}

\item{K}{TODO: check what this does (number of iterations for plugin lasso, probably).}

\item{verbose}{If TRUE, prints information to the screen while evaluating.}

\item{lambda}{Penalty parameter (a number).}

\item{phipost}{TODO: check this (something with glmnet).}
}
\value{
A list with (TODO).
}
\description{
TODO.
}
