% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xvalidate.R
\name{select_fes}
\alias{select_fes}
\title{Title}
\usage{
select_fes(fe_list, select_obs, list = TRUE)
}
\arguments{
\item{fe_list}{A list of fixed effects.}

\item{select_obs}{A vector of selected observations / rows.}

\item{list}{Logical (TODO: check what this does).}
}
\value{
A modified list of fixed effects.
}
\description{
A helper function for \code{xvalidate} (TODO: add some description of what this does).
}
\examples{
#TODO: add some examples.
}
