% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_wt}
\alias{standardize_wt}
\title{Weighted Standardization}
\usage{
standardize_wt(x, weights = rep(1/n, n), intercept = TRUE, return.sd = FALSE)
}
\arguments{
\item{x}{Regressor matrix.}

\item{weights}{Weights.}

\item{intercept}{Logical. If \code{TRUE}, adds an intercept.}

\item{return.sd}{Logical. If \code{TRUE}, returns also standard errors.}
}
\value{
TODO: check what this gives.
}
\description{
Performs weighted standardization of x variables. Needed for penalized GLM.
}
\examples{
# TODO: add examples here.
}
