% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xeex}
\alias{xeex}
\title{XeeX Matrix Computation}
\usage{
xeex(X, ee)
}
\arguments{
\item{X, ee}{Matrices.}
}
\value{
The matrix product X(k)'ee'X(k).
}
\description{
Given matrix ee' and matrix X, compute X(k)'ee'X(k) for each regressor X.
}
