% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xvalidate.R
\name{xvalidate}
\alias{xvalidate}
\title{Implementing Cross Validation}
\usage{
xvalidate(
  y,
  x,
  fes,
  IDs,
  testID = NULL,
  tol = 1e-08,
  hdfetol = 1e-04,
  colcheck = TRUE,
  init_mu = NULL,
  init_x = NULL,
  init_z = NULL,
  verbose = FALSE,
  folds = NULL,
  cluster = NULL,
  penalty = c("none", "lasso", "ridge"),
  method = "placeholder",
  standardize = TRUE,
  penweights = rep(1, ncol(x_reg)),
  lambda = 0
)
}
\arguments{
\item{y}{Dependent variable (a vector)}

\item{x}{Regressor matrix.}

\item{fes}{List of fixed effects.}

\item{IDs}{TODO: check what this does (probably for cross validation).}

\item{testID}{TODO: check this.}

\item{tol}{Tolerance parameter for convergence of the IRLS algorithm.}

\item{hdfetol}{Tolerance parameter for the within-transformation step,
passed on to \code{lfe::demeanlist}.}

\item{colcheck}{Logical. If \code{TRUE}, checks for perfect multicollinearity in \code{x}.}

\item{init_mu}{TODO: check this.}

\item{init_x}{TODO: check this.}

\item{init_z}{TODO: check this.}

\item{verbose}{Logical. If \code{TRUE}, it prints information to the screen while evaluating.}

\item{folds}{TODO: check this.}

\item{cluster}{Optional: a vector classifying observations into clusters (to use when calculating SEs).}

\item{penalty}{A string indicating the penalty type. Currently supported: "lasso" and "ridge".}

\item{method}{The user can set this equal to "plugin" to perform the plugin algorithm with
coefficient-specific penalty weights (see details). Otherwise, a single global penalty is used.}

\item{standardize}{Logical. If \code{TRUE}, x variables are standardized before estimation.}

\item{penweights}{Optional: a vector of coefficient-specific penalties to use in plugin lasso when
\code{method == "plugin"}.}

\item{lambda}{Penalty parameter, to be passed on to penhdfeppml_int or penhdfeppml_cluster_int.}
}
\value{
A list (TODO: check this).
}
\description{
TODO: put some description here.
}
\examples{
#TODO: add examples.
}
