% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cluster_matrix}
\alias{cluster_matrix}
\title{Cluster Standard Error EEstimation}
\usage{
cluster_matrix(e, cluster, x)
}
\arguments{
\item{e}{Vector of residuals.}

\item{cluster}{Vector of clusters.}

\item{x}{Regressor matrix.}
}
\value{
Gives the XeeX matrix (TODO: check this).
}
\description{
TODO: add explanation here.
}
\examples{
# TODO: add examples here.
}
