% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{Country ISO Codes}
\format{
A data frame with 249 rows and 4 variables.
\itemize{
\item iso: Country ISO 3166 code.
\item name: Country name.
\item region: Continent.
\item subregion: sub-continental region.
}
}
\source{
The source of the data set is Luke Duncalfe's ISO-3166-Countries-with-Regional-Codes repository
on GitHub (\url{https://github.com/lukes/ISO-3166-Countries-with-Regional-Codes#readme}).
}
\usage{
countries
}
\description{
An auxiliary data set with basic geographic information about country ISO 3166 codes included in the
\code{trade} data set.
}
\keyword{datasets}
