% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{genfes}
\alias{genfes}
\title{Generating a List of Fixed Effects}
\usage{
genfes(data, f1, f2)
}
\arguments{
\item{data}{A data frame including the factors.}

\item{f1}{A character vector with the names of variables to serve as first term in the interaction.}

\item{f2}{A character vector with the names of variables (second term).}
}
\value{
A list containing the interacted factors.
}
\description{
\code{genfes} generates a list of fixed effects by creating interactions of paired factors.
}
\examples{
data <- data.frame(exporter = c("Spain", "Spain", "Portugal", "France"),
                   importer = c("China", "Swaziland", "Mozambique", "Tunisia"),
                   year = c("1998", "1999", "1998", "1999"))
genfes(data, f1 = c("exporter",  "importer", "exporter"), f2 = c("year",  "year", "importer"))
}
