% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{genfes}
\alias{genfes}
\title{Generating a List of Fixed Effects}
\usage{
genfes(data, inter)
}
\arguments{
\item{data}{A data frame including the factors.}

\item{inter}{A list: each element includes the variables to be interacted (both names and column}
}
\value{
A list containing the desired interactions of \code{vars}, with the same length as \code{inter}.
}
\description{
\code{genfes} generates a list of fixed effects by creating interactions of paired factors.
}
