% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdfeppml.R
\name{hdfeppml}
\alias{hdfeppml}
\title{Poisson Pseudo Maximum Likelihood Estimation}
\usage{
hdfeppml(
  y,
  x,
  fes,
  tol = 1e-08,
  hdfetol = 1e-04,
  colcheck = TRUE,
  selectobs = NULL,
  mu = NULL,
  saveX = TRUE,
  init_z = NULL,
  verbose = FALSE,
  maxiter = 1000,
  cluster = NULL,
  vcv = TRUE
)
}
\arguments{
\item{y}{Dependent variable (a vector)}

\item{x}{Regressor matrix.}

\item{fes}{List of fixed effects.}

\item{tol}{Tolerance parameter.}

\item{hdfetol}{Tolerance parameter for fixed effects, passed on to \code{lfe::demeanlist}.}

\item{colcheck}{Logical. If \code{TRUE}, checks for perfect multicollinearity in \code{x}.}

\item{selectobs}{A numeric vector with selected observations / rows (optional).}

\item{mu}{TODO: check this.}

\item{saveX}{Logical. If \code{TRUE}, it saves the residuals of the model.}

\item{init_z}{TODO: check this.}

\item{verbose}{If TRUE, prints information to the screen while evaluating.}

\item{maxiter}{Maximum number of iterations (a number).}

\item{cluster}{A vector classifying observations into clusters (TODO: check).}

\item{vcv}{TODO: check this. Something to do with standard errors.}
}
\value{
A list (TODO: complete this).
}
\description{
\code{hdfeppml} implements (unpenalized) PPML estimation in a setting with high-dimensional
fixed effects.
}
\examples{
y <- trade$export
x <- data.matrix(trade[,-1:-9])
fes <- genfes(trade,
              f1 = c("exp",  "imp", "exp"),
              f2 = c("time",  "time", "imp"))
reg <- hdfeppml(y = y, x = y, fes = fes)
}
