% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdfeppml.R
\name{hdfeppml}
\alias{hdfeppml}
\title{Poisson Pseudo Maximum Likelihood Estimation}
\usage{
hdfeppml(
  y,
  x,
  fes,
  tol = 1e-08,
  hdfetol = 1e-04,
  colcheck = TRUE,
  selectobs = NULL,
  mu = NULL,
  saveX = TRUE,
  init_z = NULL,
  verbose = FALSE,
  maxiter = 1000,
  cluster = NULL,
  vcv = TRUE
)
}
\arguments{
\item{y}{Dependent variable (a vector)}

\item{x}{Regressor matrix.}

\item{fes}{List of fixed effects.}

\item{tol}{Tolerance parameter.}

\item{hdfetol}{Tolerance parameter for fixed effects, passed on to \code{lfe::demeanlist}.}

\item{colcheck}{Logical. If \code{TRUE}, checks for perfect multicollinearity in \code{x}.}

\item{selectobs}{A numeric vector with selected observations / rows (optional).}

\item{mu}{Optional: initial values of the \eqn{\mu} "weights",  to be used in the
first iteration of the algorithm.}

\item{saveX}{Logical. If \code{TRUE}, it returns the values of x and z after partialling out the
fixed effects.}

\item{init_z}{Optional: initial values of the transformed dependent variable, to be used in the
first iteration of the algorithm.}

\item{verbose}{If TRUE, prints information to the screen while evaluating.}

\item{maxiter}{Maximum number of iterations (a number).}

\item{cluster}{A vector classifying observations into clusters.}

\item{vcv}{TODO: check this. Something to do with standard errors.}
}
\value{
A list with the following elements:
\itemize{
\item \code{coefficients}: coefficient (beta) estimates.
\item \code{residuals}: residuals of the model.
\item \code{mu}: (TODO: check this)
\item \code{deviance}: (TODO: check this)
\item \code{bic}: (TODO: check this)
\item \code{x_resid}: matrix of demeaned regressors.
\item \code{z_resid}: vector of demeaned (transformed) dependent variable.
\item \code{se}: standard errors of the coefficients.
}
}
\description{
\code{hdfeppml} implements (unpenalized) PPML estimation in the presence of high-dimensional
fixed effects.
}
\details{
Internally, \code{hdfeppml} performs iteratively re-weighted least squares (IRLS) on a transformed
model, as described in Breinlich, Corradi, Rocha, Ruta, Santos Silva and Zylkin (2021). In each
iteration, the function calculates the transformed dependent variable, partials out the fixed effects
(calling \code{lfe::demeanlist}) and then solves a weighted least squares problem (using fast C++
implementation).
}
\examples{
y <- trade$export
x <- data.matrix(trade[,-1:-9])
fes <- list(exp_time = interaction(trade$exp, trade$time),
            imp_time = interaction(trade$imp, trade$time),
            pair     = interaction(trade$exp, trade$imp))
reg <- hdfeppml(y = y, x = x, fes = fes)
}
