% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlfitppml.R
\name{mlfitppml}
\alias{mlfitppml}
\title{General Penalized PPML Estimation}
\usage{
mlfitppml(
  y,
  x,
  fes,
  lambdas,
  penalty = "lasso",
  tol = 1e-08,
  hdfetol = 1e-04,
  colcheck = TRUE,
  post = TRUE,
  cluster = NULL,
  method = "bic",
  IDs = 1:n,
  verbose = FALSE,
  xval = FALSE,
  standardize = TRUE,
  vcv = TRUE,
  penweights = NULL,
  K = 15
)
}
\arguments{
\item{y}{Dependent variable (a vector)}

\item{x}{Regressor matrix.}

\item{fes}{List of fixed effects.}

\item{lambdas}{Vector of penalty parameters.}

\item{penalty}{A string. Currently supported: "lasso", "ridge", "SCAD".}

\item{tol}{Tolerance parameter.}

\item{hdfetol}{Tolerance parameter for fixed effects, passed on to \code{lfe::demeanlist}.}

\item{colcheck}{Logical. If \code{TRUE}, checks for perfect multicollinearity in \code{x}.}

\item{post}{Logical. If \code{TRUE}, estimates a post-penalty model with the selected variables.}

\item{cluster}{A vector classifying observations into clusters.}

\item{method}{TODO: check what this does.}

\item{IDs}{TODO: check what this does (probably for cross validation).}

\item{verbose}{If TRUE, prints information to the screen while evaluating.}

\item{xval}{Logical. If \code{TRUE}, carries out cross-validation.}

\item{standardize}{Logical. If \code{TRUE}, x variables are standardized (TODO: check this).}

\item{vcv}{TODO: check this.}

\item{penweights}{TODO: check what this does (parameter-specific penalties in plugin lasso?).}

\item{K}{TODO: check what this does.}
}
\value{
A list (TODO: complete this).
}
\description{
\code{mlfitppml} is a general wrapper function for penalized PPML estimation, calling
\code{penhdfeppml}, \code{penhdfeppml_cluster} and \code{hdfeppml} as needed.
}
\examples{
# TODO: add examples here.
}
