% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penhdfeppml.R
\name{penhdfeppml}
\alias{penhdfeppml}
\title{Title}
\usage{
penhdfeppml(
  y,
  x,
  fes,
  lambda,
  tol = 1e-08,
  hdfetol = 1e-04,
  glmnettol = 1e-12,
  penalty = c("lasso", "ridge"),
  penweights = NULL,
  selectobs = rep(TRUE, length(y)),
  saveX = TRUE,
  mu = NULL,
  colcheck = TRUE,
  init_z = NULL,
  post = FALSE,
  verbose = FALSE,
  standardize = TRUE,
  method = "placeholder",
  cluster = NULL,
  debug = FALSE
)
}
\arguments{
\item{debug}{}
}
\value{

}
\description{
Title
}
