% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penhdfeppml_cluster.R
\name{penhdfeppml_cluster}
\alias{penhdfeppml_cluster}
\title{Title}
\usage{
penhdfeppml_cluster(
  y,
  x,
  fes,
  cluster,
  tol = 1e-08,
  hdfetol = 1e-04,
  glmnettol = 1e-12,
  penalty = "lasso",
  penweights = NULL,
  selectobs = NULL,
  saveX = TRUE,
  mu = NULL,
  colcheck = TRUE,
  K = 15,
  init_z = NULL,
  post = FALSE,
  verbose = FALSE,
  lambda = NULL
)
}
\arguments{
\item{lambda}{}
}
\value{

}
\description{
Title
}
