% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penhdfeppml_cluster.R
\name{penhdfeppml_cluster}
\alias{penhdfeppml_cluster}
\title{Plugin Lasso Estimation}
\usage{
penhdfeppml_cluster(
  y,
  x,
  fes,
  cluster,
  tol = 1e-08,
  hdfetol = 1e-04,
  glmnettol = 1e-12,
  penalty = "lasso",
  penweights = NULL,
  selectobs = NULL,
  saveX = TRUE,
  mu = NULL,
  colcheck = TRUE,
  K = 15,
  init_z = NULL,
  post = FALSE,
  verbose = FALSE,
  lambda = NULL
)
}
\arguments{
\item{y}{Dependent variable (a vector)}

\item{x}{Regressor matrix.}

\item{fes}{List of fixed effects.}

\item{cluster}{A vector classifying observations into clusters.}

\item{tol}{Tolerance parameter.}

\item{hdfetol}{Tolerance parameter for fixed effects, passed on to \code{lfe::demeanlist}.}

\item{glmnettol}{Tolerance parameter to be passed on to glmnet.}

\item{penalty}{A string. Currently supported: "lasso", "ridge", "SCAD".}

\item{penweights}{TODO: check what this does (parameter-specific penalties in plugin lasso?).}

\item{selectobs}{A numeric vector with selected observations / rows (optional).}

\item{saveX}{Logical. If \code{TRUE}, it returns the values of x and z after partialling out the
fixed effects.}

\item{mu}{Optional: initial values of the \eqn{\mu} "weights",  to be used in the
first iteration of the algorithm.}

\item{colcheck}{Logical. If \code{TRUE}, checks for perfect multicollinearity in \code{x}.}

\item{K}{TODO: check what this does.}

\item{init_z}{Optional: initial values of the transformed dependent variable, to be used in the
first iteration of the algorithm.}

\item{post}{Logical. If \code{TRUE}, estimates a post-penalty model with the selected variables.}

\item{verbose}{If TRUE, prints information to the screen while evaluating.}

\item{lambda}{Penalty parameter (a number).}
}
\value{
TODO: check this.
}
\description{
Estimates coefficient-specific penalty weights that account for heteroskedasticity.
Called by \code{mlfitppml} and \code{penhdfeppml}.
}
\examples{
# TODO: add examples here.
}
