% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xvalidate.R
\name{select_fes}
\alias{select_fes}
\title{Filtering fixed effect lists}
\usage{
select_fes(fe_list, select_obs, list = TRUE)
}
\arguments{
\item{fe_list}{A list of fixed effects.}

\item{select_obs}{A vector of selected observations / rows.}

\item{list}{Logical. If \code{TRUE}, it returns a list. Otherwise, a data frame.}
}
\value{
A modified list of fixed effects.
}
\description{
A helper function for \code{xvalidate} that filters a list of fixed effects and returns the modified
list. Used to split the fixed effects for cross-validation.
}
