% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_wt}
\alias{standardize_wt}
\title{Weighted Standardization}
\usage{
standardize_wt(x, weights = rep(1/n, n), intercept = TRUE, return.sd = FALSE)
}
\arguments{
\item{x}{Regressor matrix.}

\item{weights}{Weights.}

\item{intercept}{Logical. If \code{TRUE}, adds an intercept.}

\item{return.sd}{Logical. If \code{TRUE}, it returns standard errors for the means.}
}
\value{
If \code{return.sd == FALSE}, it gives the matrix of standardized regressors. If
\code{return.sd == TRUE}, then it returns the vector of standard errors of the means of the
variables.
}
\description{
Performs weighted standardization of x variables. Used in \code{fastridge}.
}
