% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{wrappers}
\alias{wrappers}
\alias{mlfitppml2}
\alias{hdfeppml2}
\alias{penhdfeppml2}
\alias{penhdfeppml_cluster2}
\title{PPML Estimation for Data Frames}
\usage{
mlfitppml2(data, dep = 1, indep = NULL, fixed = NULL, interactions = NULL, ...)

hdfeppml2(data, dep = 1, indep = NULL, fixed = NULL, interactions = NULL, ...)

penhdfeppml2(
  data,
  dep = 1,
  indep = NULL,
  fixed = NULL,
  interactions = NULL,
  ...
)

penhdfeppml_cluster2(
  data,
  dep = 1,
  indep = NULL,
  fixed = NULL,
  interactions = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame containing all relevant variables.}

\item{dep}{A string with the name of the independent variable or a column number.}

\item{indep}{A vector with the names or column numbers of the regressors. If left unspecified,
all remaining variables (excluding fixed effects) are included in the regressor matrix.}

\item{fixed}{A vector with the names or column numbers of factor variables identifying the fixed effects.}

\item{interactions}{A list with the desired interactions between the variables in \code{fixed}.
Optional: if left unspecified, the function will just use the variables in \code{fixed}.}

\item{...}{Further arguments, to be passed on to the main function.}
}
\value{
The same as their respective parent functions.
}
\description{
These functions are thin wrappers around \code{mlfitppml}, \code{hdfeppml}, \code{penhdfeppml} and
\code{penhdfeppml_cluster}, providing a more convenient interface for data frames. Whereas the original
functions require some preliminary handling of data sets (\code{y} must be a vector, \code{x} must be a
matrix and \code{fes} must be provided in a list), the wrappers take a full data frame in the \code{data}
argument, and users can simply specify which variables correspond to y, x and the fixed effects, using
either variable names or column numbers.
}
\examples{
test1 <- mlfitppml2(data = trade[, -(5:10)],
                    dep = "export",
                    fixed = c("exp", "imp", "time"),
                    interactions = list(c("exp", "time"),
                                        c("imp", "time"),
                                        c("exp", "imp")),
                    lambdas = c(0.05, 0.01, 0.005, 0.001, 0.0005, 0.0001))
test2 <- hdfeppml2(data = trade[, -(5:10)],
                   dep = "export",
                   fixed = c("exp", "imp", "time"),
                   interactions = list(c("exp", "time"),
                                       c("imp", "time"),
                                       c("exp", "imp")))
test3 <- penhdfeppml2(data = trade[, -(5:10)],
                      dep = "export",
                      fixed = c("exp", "imp", "time"),
                      interactions = list(c("exp", "time"),
                                          c("imp", "time"),
                                          c("exp", "imp")),
                      lambda = 0.05)
test4 <- penhdfeppml_cluster2(data = trade[, -(5:10)],
                              dep = "export",
                              fixed = c("exp", "imp", "time"),
                              interactions = list(c("exp", "time"),
                                                  c("imp", "time"),
                                                  c("exp", "imp")),
                              cluster = interaction(trade$exp, trade$imp))

}
\seealso{
\code{\link[=mlfitppml]{mlfitppml()}}
}
