% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xvalidate.R
\name{xvalidate}
\alias{xvalidate}
\title{Implementing Cross Validation}
\usage{
xvalidate(
  y,
  x,
  fes,
  IDs,
  testID = NULL,
  tol = 1e-08,
  hdfetol = 1e-04,
  colcheck = TRUE,
  init_mu = NULL,
  init_x = NULL,
  init_z = NULL,
  verbose = FALSE,
  folds = NULL,
  cluster = NULL,
  penalty = c("none", "lasso", "ridge"),
  method = "placeholder",
  standardize = TRUE,
  penweights = rep(1, ncol(x_reg)),
  lambda = 0
)
}
\arguments{
\item{y}{Dependent variable (a vector)}

\item{x}{Regressor matrix.}

\item{fes}{List of fixed effects.}

\item{IDs}{TODO: check what this does (probably for cross validation).}

\item{testID}{TODO: check this.}

\item{tol}{Tolerance parameter.}

\item{hdfetol}{Tolerance parameter for fixed effects, passed on to \code{lfe::demeanlist}.}

\item{colcheck}{Logical. If \code{TRUE}, checks for perfect multicollinearity in \code{x}.}

\item{init_mu}{TODO: check this.}

\item{init_x}{TODO: check this.}

\item{init_z}{TODO: check this.}

\item{verbose}{If TRUE, prints information to the screen while evaluating.}

\item{folds}{TODO: check this.}

\item{cluster}{A vector classifying observations into clusters.}

\item{penalty}{A string. Currently supported: "lasso" and "ridge".}

\item{method}{TODO: check what this does.}

\item{standardize}{Logical. If \code{TRUE}, x variables are standardized (TODO: check this).}

\item{penweights}{TODO: check what this does (parameter-specific penalties in plugin lasso?).}

\item{lambda}{Penalty parameter, to be passed on to penhdfeppml or penhdfeppml_cluster.}
}
\value{
A list (TODO: check this).
}
\description{
TODO: put some description here.
}
\examples{
#TODO: add examples.
}
